feat: remove posts
