+++ src/notify.rs
@@ -1,7 +1,7 @@
 use std::collections::HashSet;
 
 use crate::{
-    get_subscribers, rfp::get_subscribers as get_rfp_subscribers, PostId, Proposal, ProposalId, RFP,
+    get_subscribers, rfp::get_subscribers as get_rfp_subscribers, Proposal, ProposalId, RFP,
 };
 use devhub_common::social_db_contract;
 use near_sdk::serde_json::json;
@@ -102,31 +102,6 @@ pub fn notify_rfp_subscribers(rfp: &RFP, additional_accounts: HashSet<AccountId>
     )
 }
 
-pub fn notify_mentions(text: &str, post_id: PostId) -> Promise {
-    let mentions = get_text_mentions(text);
-
-    notify_accounts(
-        env::predecessor_account_id(),
-        mentions,
-        json!({
-            "type": "devgovgigs/mention",
-            "post": post_id,
-        }),
-    )
-}
-
-pub fn notify_like(post_id: PostId, post_author: AccountId) -> Promise {
-    notify(env::predecessor_account_id(), post_author, notify_value(post_id, "like"))
-}
-
-pub fn notify_reply(post_id: PostId, post_author: AccountId) -> Promise {
-    notify(env::predecessor_account_id(), post_author, notify_value(post_id, "reply"))
-}
-
-pub fn notify_edit(post_id: PostId, post_author: AccountId) -> Promise {
-    notify(env::predecessor_account_id(), post_author, notify_value(post_id, "edit"))
-}
-
 pub fn notify_edit_proposal(proposal_id: ProposalId, post_author: AccountId) -> Promise {
     notify(
         env::current_account_id(),
@@ -138,13 +113,6 @@ pub fn notify_edit_proposal(proposal_id: ProposalId, post_author: AccountId) -> 
             "notifier": env::predecessor_account_id(),
         }),
     )
-}
-
-fn notify_value(post_id: PostId, action: &str) -> serde_json::Value {
-    json!({
-        "type": format!("devgovgigs/{}", action),
-        "post": post_id,
-    })
 }
 
 fn notify(notifier: AccountId, post_author: AccountId, notify_value: serde_json::Value) -> Promise {
@@ -161,48 +129,4 @@ fn notify(notifier: AccountId, post_author: AccountId, notify_value: serde_json:
                 }
             }
         }))
-}
-
-#[cfg(all(test, not(target_arch = "wasm32")))]
-mod tests {
-    use super::notify_mentions;
-
-    use near_sdk::test_utils::{get_created_receipts, VMContextBuilder};
-    use near_sdk::{testing_env, VMContext};
-
-    fn get_context(is_view: bool) -> VMContext {
-        VMContextBuilder::new()
-            .signer_account_id("bob.near".parse().unwrap())
-            .is_view(is_view)
-            .build()
-    }
-
-    #[test]
-    pub fn test_notify_mentions() {
-        let context = get_context(false);
-        testing_env!(context);
-        let text = "Mentioning @a.near and @bcdefg.near";
-        notify_mentions(text, 2);
-        let receipts = get_created_receipts();
-        assert_eq!(1, receipts.len());
-
-        if let near_sdk::mock::MockAction::FunctionCallWeight { method_name, args, .. } =
-            &receipts[0].actions[0]
-        {
-            assert_eq!(method_name, b"set");
-            assert_eq!(args, b"{\"data\":{\"bob.near\":{\"index\":{\"notify\":\"[{\\\"key\\\":\\\"a.near\\\",\\\"value\\\":{\\\"type\\\":\\\"devgovgigs/mention\\\",\\\"post\\\":2}},{\\\"key\\\":\\\"bcdefg.near\\\",\\\"value\\\":{\\\"type\\\":\\\"devgovgigs/mention\\\",\\\"post\\\":2}}]\"}}}}");
-        } else {
-            assert!(false, "Expected a function call ...")
-        }
-    }
-
-    #[test]
-    pub fn test_no_mentions() {
-        let context = get_context(false);
-        testing_env!(context);
-        let text = "Not mentioning anyone";
-        notify_mentions(text, 2);
-        assert_eq!(1, get_created_receipts().len());
-        assert_eq!(0, get_created_receipts()[0].actions.len());
-    }
 }
